+++ packages/frontend/src/navigation/NavItem.tsx
@@ -13,7 +13,7 @@ const NavItem = ({ icon, description, to }: NavProps): JSX.Element => {
     <NavLink
       to={to}
       className={(isActive): string =>
-        `relative px-4 py-1 cursor-pointer mr-[1px] no-underline flex items-center text-black ${
+        `relative px-4 py-1 cursor-pointer no-underline flex items-center text-black ${
           isActive ? ' bg-gray-200' : ' hover:bg-gray-100'
         }`
       }